페이지 범위 밖이면 가장 가까운 페이지로 리다이렉트하는 컴포넌트 작성
