+++ src/pages/purchased/main/index.jsx
@@ -3,13 +3,14 @@ import { jsx } from '@emotion/core';
 import React from 'react';
 import { Helmet } from 'react-helmet';
 import { connect } from 'react-redux';
-import { Link, Redirect } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 import { ButtonType } from '../../../components/ActionBar/constants';
 import BookDownLoader from '../../../components/BookDownLoader';
 import { Books } from '../../../components/Books';
 import Editable from '../../../components/Editable';
 import Empty from '../../../components/Empty';
 import { BookError } from '../../../components/Error';
+import PageRedirect from '../../../components/PageRedirect';
 import ResponsivePaginator from '../../../components/ResponsivePaginator';
 import SearchBar from '../../../components/SearchBar';
 import SelectShelfModal from '../../../components/SelectShelfModal';
@@ -253,28 +254,13 @@ function PurchasedMain(props) {
     return { buttonProps };
   }
 
-  let redirection = null;
-  if (totalPages > 0) {
-    const realPage = Math.max(1, Math.min(totalPages, currentPage));
-    if (currentPage !== realPage) {
-      const newUrlParams = new URLSearchParams(location.search);
-      newUrlParams.set('page', realPage);
-      const newSearch = newUrlParams.toString();
-      const to = {
-        ...location,
-        search: newSearch !== '' ? `?${newSearch}` : '',
-      };
-      redirection = <Redirect to={to} />;
-    }
-  }
-
   if (showShelves) {
     return (
       <>
         <Helmet>
           <title>모든 책 - 내 서재</title>
         </Helmet>
-        {redirection}
+        <PageRedirect currentPage={currentPage} totalPages={totalPages} />
         <SelectShelfModal onBackClick={handleShelfBackClick} onShelfSelect={handleShelfSelect} />
       </>
     );
@@ -285,7 +271,7 @@ function PurchasedMain(props) {
       <Helmet>
         <title>모든 책 - 내 서재</title>
       </Helmet>
-      {redirection}
+      <PageRedirect currentPage={currentPage} totalPages={totalPages} />
       <TabBar activeMenu={TabMenuTypes.ALL_BOOKS} />
       <Editable
         allowFixed
